refactor(tests): split fake repository snapshots
